0.1.0
Updated query.sql to include test query, updated schema.sql to have  schema of the database, updated seeds.sql to populate the tables of employee. added a connection.js file to connect to the mySQL db, with dotenv enviornment variables, updated main.js to target the exported function from connection.js and query the database with getDepartments function which awaits a promise to resolve or reject then dynamically maps each dept as a new array and returns that array then departments function uses that array for the inquirer choice options.