0.3.0
Reduced file size by moving the queryEngine to query.js and importing it to main.js where it can be called with the query as an arg. Updated data to dynamically populate inquierer feilds. updated all files to have fixed indentation and use best possible practices, updated main.js to have addRole function which takes the data returned from viewAllDepts and maps it into an array of objects to be used by the inquirer line for adding a new role, then handleaddrole accepts the data from addrole by destructuring, then parses department split by - to ensure then input only returns the index of the selected vaule which is then used by the query string in a template litteral and then is called with query engine as an arg.