0.4.0
added in an addEmploye and handleAddEmployee function that uses the instance of viewallroles() and maps the selected data, in roleArr then roleArr dynamically generates choices from the db, mgmt does similar but pulls employees from managment roles and then maps into an array for dynamicly rendered choices. then data is called which handleAddEmploye function which destructures the name variable into first last role manager and parses and splits by - to get the number that remains which is assigned to a variable and used in the query string with a turnary opperator, that allows managerIndex to be NULL if no manager is selected, OR to include the values if a manager is selected. Then the query is called with queryEngine imported from query.js