1.0.1
fixed bug where console.table would be called when selecting to add since, the inquirer used dynamically populated choices from other roles useing the queryEngine method from parent class, which had a console.table statement inside of it, removed that statment and chained a .then() on inital prompt to console.log the data that is returned by the methods. Updated seeds file to better follow the ideal of the correct manager to the correct employe role. Updated to have Professional read me that inclueds description video link, license and so on.